Put compute_transition_probabilities_sparse in freestyle_solution
